I implemented a system of annotation for buffers. With this, RayPlatform knows which buffer is dirty (possibly available, but maybe not) and which buffer is available.
This change makes the RayPlatform code more compliant with
the MPI standard as a buffer given to MPI_Isend should be reused
if MPI_Test still says that the request has not completed.

Usually, the code was working because the ring buffer protected
the buffers. But it seems that when a message is sent from MPI rank
A to MPI rank A, there is no copy of it done by the MPI
library on Cray systems or when using QLogic Performance Scaled
Messaging. Therefore, in these cases, the RayPlatform code
was probably causing faults.

This may fix some issues with QLogic, Inc. Performance Scaled
Messaging and issues with Cray systems.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>