The latency on a IBM iDataPlex (guillimin at McGill) for a Ray job of 36 cores was reduced from 23 to 17 microseconds (back and forth).
The optimization is just to poll at most <number of MPI ranks> instead
of at most 1 MPI rank in the round-robin reception. If no message is pumped
in the first probe, the whole tick of the RayPlatform virtual
machine will be wasted anyway because the slave mode is probably waiting
for a reply.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>