MessageRouter: store the routing information in the buffer
To route a message, we need to store its true source and its true
destination somewhere. The initial version of the virtual router stored
this information in the MPI tag itself.

According to the MPI standard, MPI tags must be greater than 0 and at
most MPI_TAG_UB. The MPI standard says the MPI_TAG_UB is at least 32767
(2^15 - 1). Regardless, RayPlatform needs to know if a message is a
routed message or not. To do so, RayPlatform requires the application
to use tag values from 0 to 16383 (2^14 - 1). If a message is routed,
its tag is TRUE_TAG + 16384. This can be checked with tag >= 16384.
This is compliant with the MPI standard, while the old code was not.

The true source and the true destination are now stored in the last
message unit of the buffer. The maximum number of message units is 500
for the application, but there is a true maximum of 501. The last one
is reserved for RayPlatform.

Finally, if a message has no buffer, one is allocated. With this, there
is no limit for the number of processes. Before this patch, the maximum
number of processes with the virtual router was 2048 with Open-MPI.

Before this patch, the MPI tag was limited to 8 bits (0-255) because
routing information was stored in the tag as well. The new range for
MPI tags with the virtual router is 0-16385.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>