ComputeCore: routed messages must be purged
If a message is routed, the original message needs to be destroyed
with fire. Otherwise, a slave mode or a master mode may pick it up,
which would cause silly behaviors.

This bug only occurred with the new message router and only
messages picked up by a slave mode or a master mode were possibly
causing problems. Routed messages could not be wrongly handled by
the message tag interruption table.

Reviewed-by: Élénie Godzaridis <elenie.godzaridis@bentley.com>
Tested-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>
Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>