core: The old behavior (no mini-ranks) now works as expected
The implementation of mini-ranks added some complexity. But this
complexity was reduced and now the same code runs in pure MPI
mode (1 UNIX process per MPI rank) or in mini-ranks mode (with
1 IEEE POSIX thread per mini-rank). If the end user does not
say anything about mini-ranks, the default is to use 1 process
per MPI rank and no thread at all (and no mini-rank).

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>