Message reception is now interleaved with send operations.
The obvious way to implement the communication layer is to receive
and send messages in a loop (receive followed by send). However,
although receiving a message is always fast, sending messages is
slower because there is a loop over all the mini-ranks. This patch
fixes this by interleaving the reception in the sending loop.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>