communication: relay buffer bytes instead of buffer 64-bit integers
This fixes broken message routing code for the actor model.

The option -route-messages was working as usual, but any code
that was using the actor model could not use -route-messages.

This was fixed. The problem was due to the relay method in the message
router that was incorrectly forwarding routed messages to their
destination.

Command tested with 2.3.0 (works):
   mpiexec -n 25 ./Ray -route-messages -test-network-only -output LOL

Link: http://en.wikipedia.org/wiki/Actor_model
Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>